requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
idna==2.7
python-dotenv==0.9.1
requests==2.20.0
spotipy==2.4.4
urllib3==1.24.1
